Correção de float para double

+++ Projeto2/src/Model/Produto/Eletrodomesticos.java
@@ -11,7 +11,7 @@ Willian Yoshio Murayama
  */
 public class Eletrodomesticos extends Produto implements Serializable {
 
-    public Eletrodomesticos(int codigo, String nome, String descricao, LocalDate dataFabricacao, float valor, Fabricante fabricante, boolean disponivel) {
+    public Eletrodomesticos(int codigo, String nome, String descricao, LocalDate dataFabricacao, double valor, Fabricante fabricante, boolean disponivel) {
         super(codigo, nome, descricao, dataFabricacao, valor, fabricante, disponivel);
     }
